(refactor) Toolkit: Add ApplicationAdapter and remove Application parameters in favor of ApplicationFixture.
