+++ subprojects/testfx-core/src/main/java/org/testfx/api/FxToolkit.java
@@ -35,6 +35,7 @@ import org.testfx.toolkit.ApplicationFixture;
 import org.testfx.toolkit.ApplicationLauncher;
 import org.testfx.toolkit.ApplicationService;
 import org.testfx.toolkit.ToolkitService;
+import org.testfx.toolkit.impl.ApplicationAdapter;
 import org.testfx.toolkit.impl.ApplicationLauncherImpl;
 import org.testfx.toolkit.impl.ApplicationServiceImpl;
 import org.testfx.toolkit.impl.ToolkitServiceImpl;
@@ -170,28 +171,25 @@ public class FxToolkit {
         );
     }
 
-    @Unstable(reason = "is missing apidocs")
-    public static Application setupApplication(Application application,
-                                               String... applicationArgs)
-                                        throws TimeoutException {
-        return waitForSetup(
-            service.setupApplication(
-                () -> context.getRegisteredStage(),
-                application,
-                applicationArgs
-            )
-        );
-    }
+    //@Unstable(reason = "is missing apidocs")
+    //public static Application setupApplication(Application application,
+    //                                           String... applicationArgs)
+    //                                    throws TimeoutException {
+    //    return waitForSetup(
+    //        service.setupApplication(
+    //            () -> context.getRegisteredStage(),
+    //            new ApplicationAdapter(application)
+    //        )
+    //    );
+    //}
 
     @Unstable(reason = "is missing apidocs")
-    public static ApplicationFixture setupApplication(ApplicationFixture applicationFixture,
-                                                      String... applicationArgs)
+    public static ApplicationFixture setupApplication(ApplicationFixture applicationFixture)
                                                throws TimeoutException {
         return waitForSetup(
             service.setupApplication(
                 () -> context.getRegisteredStage(),
-                applicationFixture,
-                applicationArgs
+                applicationFixture
             )
         );
     }
@@ -200,7 +198,7 @@ public class FxToolkit {
     public static void cleanupApplication(Application application)
                                    throws TimeoutException {
         waitForSetup(
-            service.cleanupApplication(application)
+            service.cleanupApplication(new ApplicationAdapter(application))
         );
     }
